Find the "descendants" of a given capture history

Given any encoded capture history, find all the encoded capture histories that include the original capture history and any other lists

Usage

descendants(k, nlists, omitk = FALSE)

Arguments

k

An encoded capture history

nlists

The total number of lists

omitk

Determine whether the original capture history is included as a descendant of itself. If omitk=TRUE it is not.

Value

a vector giving the encoded versions of the descendants

References

Silverman, B. W., Chan, L. and Vincent, K., (2024). Bootstrapping Multiple Systems Estimates to Account for Model Selection Statistics and Computing, 34(44), Available from \doi{10.1007/s11222-023-10346-9}.

Examples

descendants(2,5)
#>  [1]  2  4  6  8 10 12 14 16 18 20 22 24 26 28 30 32
descendants(5,10)
#>   [1]    5    6    7    8   13   14   15   16   21   22   23   24   29   30   31
#>  [16]   32   37   38   39   40   45   46   47   48   53   54   55   56   61   62
#>  [31]   63   64   69   70   71   72   77   78   79   80   85   86   87   88   93
#>  [46]   94   95   96  101  102  103  104  109  110  111  112  117  118  119  120
#>  [61]  125  126  127  128  133  134  135  136  141  142  143  144  149  150  151
#>  [76]  152  157  158  159  160  165  166  167  168  173  174  175  176  181  182
#>  [91]  183  184  189  190  191  192  197  198  199  200  205  206  207  208  213
#> [106]  214  215  216  221  222  223  224  229  230  231  232  237  238  239  240
#> [121]  245  246  247  248  253  254  255  256  261  262  263  264  269  270  271
#> [136]  272  277  278  279  280  285  286  287  288  293  294  295  296  301  302
#> [151]  303  304  309  310  311  312  317  318  319  320  325  326  327  328  333
#> [166]  334  335  336  341  342  343  344  349  350  351  352  357  358  359  360
#> [181]  365  366  367  368  373  374  375  376  381  382  383  384  389  390  391
#> [196]  392  397  398  399  400  405  406  407  408  413  414  415  416  421  422
#> [211]  423  424  429  430  431  432  437  438  439  440  445  446  447  448  453
#> [226]  454  455  456  461  462  463  464  469  470  471  472  477  478  479  480
#> [241]  485  486  487  488  493  494  495  496  501  502  503  504  509  510  511
#> [256]  512  517  518  519  520  525  526  527  528  533  534  535  536  541  542
#> [271]  543  544  549  550  551  552  557  558  559  560  565  566  567  568  573
#> [286]  574  575  576  581  582  583  584  589  590  591  592  597  598  599  600
#> [301]  605  606  607  608  613  614  615  616  621  622  623  624  629  630  631
#> [316]  632  637  638  639  640  645  646  647  648  653  654  655  656  661  662
#> [331]  663  664  669  670  671  672  677  678  679  680  685  686  687  688  693
#> [346]  694  695  696  701  702  703  704  709  710  711  712  717  718  719  720
#> [361]  725  726  727  728  733  734  735  736  741  742  743  744  749  750  751
#> [376]  752  757  758  759  760  765  766  767  768  773  774  775  776  781  782
#> [391]  783  784  789  790  791  792  797  798  799  800  805  806  807  808  813
#> [406]  814  815  816  821  822  823  824  829  830  831  832  837  838  839  840
#> [421]  845  846  847  848  853  854  855  856  861  862  863  864  869  870  871
#> [436]  872  877  878  879  880  885  886  887  888  893  894  895  896  901  902
#> [451]  903  904  909  910  911  912  917  918  919  920  925  926  927  928  933
#> [466]  934  935  936  941  942  943  944  949  950  951  952  957  958  959  960
#> [481]  965  966  967  968  973  974  975  976  981  982  983  984  989  990  991
#> [496]  992  997  998  999 1000 1005 1006 1007 1008 1013 1014 1015 1016 1021 1022
#> [511] 1023 1024